requirements.txt
opencv-python
numpy
pillow
ruamel.yaml==0.17.17
colorama==0.4.4; sys_platform == 'win32'
coloredlogs==15.0.1
